+++ tests/linter/rules/naming/wrong_case_in_keyword_call/test_rule.py
@@ -15,18 +15,18 @@ class TestRuleAcceptance(RuleAcceptance):
         )
 
     def test_rule_pabot(self):
-        self.check_rule(src_files=["pabot_run_keywords.robot"], expected_file=None)
+        self.check_rule(src_files=["pabot_run_keywords.robot"], expected_file="expected_output_pabot.txt")
 
     def test_convention_first_word_capitalized(self):
         self.check_rule(
-            configure=["wrong-case-in-keyword-name.convention=first_word_capitalized"],
+            configure=["wrong-case-in-keyword-call.convention=first_word_capitalized"],
             src_files=["first_word"],
             expected_file="first_word/expected_output.txt",
         )
 
     def test_configure_pattern(self):
         self.check_rule(
-            configure=[r"wrong-case-in-keyword-name.pattern=Foo\.bar"],
+            configure=[r"wrong-case-in-keyword-call.pattern=Foo\.bar"],
             src_files=["configure_pattern"],
             expected_file="configure_pattern/expected_output.txt",
         )